+++ packages/components/scripts/config.js
@@ -1,0 +1,26 @@
+/**
+ * List of components. Add a component name here to make it available to the outside
+ * @type {string[]}
+ */
+export const components = [
+  'input',
+  'button',
+];
+
+/**
+ * List of events that should be created.
+ * Add a component name here to make it available to the outside
+ * @todo: Automate this depending on components!
+ * @type {string[]}
+ */
+export const events = [
+  'sl-blur',
+  'sl-focus',
+  'sl-invalid',
+  'sl-load',
+  'sl-error',
+  'sl-blur',
+  'sl-change',
+  'sl-clear',
+  'sl-input',
+];