+++ packages/components/scripts/jobs/react/index.js
@@ -1,6 +1,4 @@
-export * from './adjustPackageVersion.js';
 export * from './createWrappers.js';
 export * from './esbuild.js';
 export * from './format.js';
-export * from './prepare.js';
 export * from './tsc.js';